SQL_Code
Tried to design program which work as a SQL works, I had only 1.5 hours to think and design the code. I did my best it was ammmmazzing experience. I will work on this further to move this code on next level.
if someone else also wants to work on it, please feel free to download the code, explore the work and try to make it more amazing.
enjoyeeeeeeeeeeee!!!!!!!